separate abstract functions from motion-csv-import
