+++ client/src/app/site/motions/models/view-csv-create-motion.ts
@@ -2,17 +2,14 @@ import { ViewCreateMotion } from './view-create-motion';
 import { CreateMotion } from './create-motion';
 
 /**
- * Interface for imported secondary data. A name can be matched to an existing
- * model instance by the solve... functions.
- * TODO MultiId will be filled if there is more than one match (to be used in case of 'I want to select one of these matches)
+ * Interface for correlating between strings representing BaseModels and existing
+ * BaseModels.
  */
 export interface CsvMapping {
     name: string;
     id?: number;
     multiId?: number[];
 }
-
-type CsvImportStatus = 'new' | 'error' | 'done';
 
 /**
  * Create motion class for the View. Its different to ViewMotion in fact that the submitter handling is different
@@ -38,66 +35,8 @@ export class ViewCsvCreateMotion extends ViewCreateMotion {
      */
     public csvSubmitters: CsvMapping[];
 
-    /**
-     * The current import status of this motion.
-     * Starts as 'new', if set to 'done', a proper {@link Motion} model will
-     * probably exist in the dataStore. error status will be set if the import
-     * cannot be done
-     */
-    private _status: CsvImportStatus = 'new';
-
-    /**
-     * list of import errors See {@link MotionImportService}
-     */
-    public errors: string[] = [];
-
-    /**
-     * Returns the current status.
-     */
-    public get status(): CsvImportStatus {
-        return this._status;
-    }
-
-    public set status(status: CsvImportStatus) {
-        this._status = status;
-    }
-
-    public get motion(): CreateMotion {
-        return this._motion;
-    }
-
     public constructor(motion?: CreateMotion) {
         super(motion);
-    }
-
-    /**
-     * Duplicate this motion into a copy of itself
-     */
-    public copy(): ViewCreateMotion {
-        return new ViewCreateMotion(
-            this._motion,
-            this._category,
-            this._submitters,
-            this._supporters,
-            this._workflow,
-            this._state
-        );
-    }
-
-    /**
-     * Checks if a given error is present. TODO: Is more a ViewModel option
-     *
-     * @param error
-     */
-    public hasError(error: string): boolean {
-        return this.errors.includes(error);
-    }
-
-    /**
-     * Toggle to set a CreateMotion to a 'successfully parsed' status
-     */
-    public done(): void {
-        this._status = 'done';
     }
 
     /**
