+++ client/src/app/site/motions/services/motion-import.service.ts
@@ -1,7 +1,6 @@
-import { BehaviorSubject, Observable } from 'rxjs';
-import { Injectable, EventEmitter } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { MatSnackBar } from '@angular/material';
-import { Papa, PapaParseConfig } from 'ngx-papaparse';
+import { Papa } from 'ngx-papaparse';
 import { TranslateService } from '@ngx-translate/core';
 
 import { Category } from 'app/shared/models/motions/category';
@@ -11,49 +10,9 @@ import { MotionBlock } from 'app/shared/models/motions/motion-block';
 import { MotionBlockRepositoryService } from './motion-block-repository.service';
 import { MotionRepositoryService } from './motion-repository.service';
 import { UserRepositoryService } from '../../users/services/user-repository.service';
-import { ViewMotion } from '../models/view-motion';
 import { ViewCsvCreateMotion, CsvMapping } from '../models/view-csv-create-motion';
-
-/**
- * Interface for value- Label combinations.
- * Map objects didn't work, TODO: Use map objects (needs iterating through all objects of a map)
- */
-export interface ValueLabelCombination {
-    value: string;
-    label: string;
-}
-
-/**
- * Interface for a new Motion and their (if any) duplicates
- */
-export interface NewMotionEntry {
-    newMotion: ViewCsvCreateMotion;
-    duplicates: ViewMotion[];
-}
-
-/**
- * interface for a preview summary
- */
-interface ImportMotionCSVPreview {
-    total: number;
-    duplicates: number;
-    errors: number;
-    new: number;
-    done: number;
-}
-
-/**
- * List of possible import errors specific for motion imports.
- */
-const errorList = {
-    MotionBlock: 'Could not resolve the motion block',
-    Category: 'Could not resolve the category',
-    Submitters: 'Could not resolve the submitters',
-    Title: 'A title is required',
-    Text: "A content in the 'text' column is required",
-    Duplicates: 'A motion with this identifier already exists.',
-    generic: 'Server upload failed' // TODO
-};
+import { TextImportService, NewEntry } from 'app/core/services/text-import.service';
+import { ViewMotion } from '../models/view-motion';
 
 /**
  * Service for motion imports
@@ -61,59 +20,19 @@ const errorList = {
 @Injectable({
     providedIn: 'root'
 })
-export class MotionImportService {
-    /** The header (order and items) that is expected from the imported file
-     *
-     */
-    public expectedHeader = [
-        'identifier',
-        'title',
-        'text',
-        'reason',
-        'submitters',
-        'category',
-        'origin',
-        'motion block'
-    ];
-
+export class MotionImportService extends TextImportService<ViewMotion> {
     /**
-     * The last parsed file object (may be reparsed with new encoding, thus kept in memory)
+     * List of possible errors and their verbose explanation
      */
-    private _rawFile: File;
-
-    /**
-     * The used column Separator. If left on an empty string (default),
-     * the papaparse parser will automatically decide on separators.
-     */
-    public columnSeparator = '';
-
-    public textSeparator = '"';
-
-    public encoding = 'utf-8';
-
-    /**
-     * List of possible encodings and their label
-     */
-    public encodings: ValueLabelCombination[] = [
-        { value: 'utf-8', label: 'UTF 8 - Unicode' },
-        { value: 'iso-8859-1', label: 'ISO 8859-1 - West European' },
-        { value: 'iso-8859-15', label: 'ISO 8859-15 - West European (with €)' }
-    ];
-
-    /**
-     * List of possible column separators
-     */
-    public columnSeparators: ValueLabelCombination[] = [
-        { label: 'Comma', value: ',' },
-        { label: 'Semicolon', value: ';' },
-        // {label: 'Tabulator', value: '\t'},
-        { label: 'Automatic', value: '' }
-    ];
-
-    public textSeparators: ValueLabelCombination[] = [
-        { label: 'Double quotes (")', value: '"' },
-        { label: "Single quotes (')", value: "'" }
-    ];
+    public errorList = {
+        MotionBlock: 'Could not resolve the motion block',
+        Category: 'Could not resolve the category',
+        Submitters: 'Could not resolve the submitters',
+        Title: 'A title is required',
+        Text: "A content in the 'text' column is required",
+        Duplicates: 'A motion with this identifier already exists.',
+        generic: 'Server upload failed' // TODO unused?
+    };
 
     /**
      * submitters that need to be created prior to importing
@@ -131,44 +50,7 @@ export class MotionImportService {
     public newMotionBlocks: CsvMapping[] = [];
 
     /**
-     * FileReader object for file import
-     */
-    private reader = new FileReader();
-
-    /**
-     * the list of parsed models that have been extracted from the opened file
-     */
-    private _entries: NewMotionEntry[] = [];
-
-    /**
-     * BehaviorSubject for displaying a preview for the currently selected entries
-     */
-    public newEntries = new BehaviorSubject<NewMotionEntry[]>([]);
-
-    /**
-     * Emits an error string to display if a file import cannot be done
-     */
-    public errorEvent = new EventEmitter<string>();
-
-    /**
-     * storing the summary preview for the import, to avoid recalculating it
-     * at each display change.
-     */
-    private _preview: ImportMotionCSVPreview;
-
-    /**
-     * Returns a summary on actions that will be taken/not taken.
-     */
-    public get summary(): ImportMotionCSVPreview {
-        if (!this._preview) {
-            this.updatePreview();
-        }
-        return this._preview;
-    }
-
-    /**
-     * Constructor. Creates a fileReader to subscribe to it for incoming parsed
-     * strings
+     * Constructor. Defines the headers expected and calls the abstract class
      * @param categoryRepo Repository to fetch pre-existing categories
      * @param motionBlockRepo Repository to fetch pre-existing motionBlocks
      * @param userRepo Repository to query/ create users
@@ -181,68 +63,61 @@ export class MotionImportService {
         private categoryRepo: CategoryRepositoryService,
         private motionBlockRepo: MotionBlockRepositoryService,
         private userRepo: UserRepositoryService,
-        private translate: TranslateService,
-        private papa: Papa,
-        private matSnackbar: MatSnackBar
+        translate: TranslateService,
+        papa: Papa,
+        matSnackbar: MatSnackBar
     ) {
-        this.reader.onload = (event: any) => {
-            // TODO type: event is a progressEvent,
-            // but has a property target.result, which typescript doesn't recognize
-            this.parseInput(event.target.result);
-        };
+        super(translate, papa, matSnackbar);
+
+        this.expectedHeader = [
+            'identifier',
+            'title',
+            'text',
+            'reason',
+            'submitters',
+            'category',
+            'origin',
+            'motion_block'
+        ];
+        this.requiredHeaderLength = 3;
     }
 
-    /**
-     * Parses the data input. Expects a string as returned by via a
-     * File.readAsText() operation
-     *
-     * @param file
-     */
-    public parseInput(file: string): void {
-        this._entries = [];
+    public clearData(): void {
         this.newSubmitters = [];
         this.newCategories = [];
         this.newMotionBlocks = [];
-        const papaConfig: PapaParseConfig = {
-            header: false,
-            skipEmptyLines: true,
-            quoteChar: this.textSeparator
+    }
+
+    /**
+     * maps an entry line to a ViewCsvCreateMotion
+     * @param line
+     */
+    public mapData(line: string): NewEntry<ViewMotion> {
+        const newEntry = new ViewCsvCreateMotion(new CreateMotion());
+        const headerLength = Math.min(this.expectedHeader.length, line.length);
+        for (let idx = 0; idx < headerLength; idx++) {
+            // iterate over items, find existing ones (their id) and collect new entries
+            switch (this.expectedHeader[idx]) {
+                case 'submitters':
+                    newEntry.csvSubmitters = this.getSubmitters(line[idx]);
+                    break;
+                case 'category':
+                    newEntry.csvCategory = this.getCategory(line[idx]);
+                    break;
+                case 'motion_block':
+                    newEntry.csvMotionblock = this.getMotionBlock(line[idx]);
+                    break;
+                default:
+                    newEntry.motion[this.expectedHeader[idx]] = line[idx];
+            }
+        }
+        const updateModels = this.repo.getMotionDuplicates(newEntry.motion);
+        return {
+            newEntry: newEntry,
+            duplicates: updateModels,
+            status: updateModels.length ? 'error' : 'new',
+            errors: updateModels.length ? ['Duplicates'] : []
         };
-        if (this.columnSeparator) {
-            papaConfig.delimiter = this.columnSeparator;
-        }
-        const entryLines = this.papa.parse(file, papaConfig).data;
-        const valid = this.checkHeader(entryLines.shift());
-        if (!valid) {
-            return;
-        }
-        entryLines.forEach(line => {
-            const newMotion = new ViewCsvCreateMotion(new CreateMotion());
-            const headerLength = Math.min(this.expectedHeader.length, line.length);
-            for (let idx = 0; idx < headerLength; idx++) {
-                // iterate over items, find existing ones (thier id) and collect new entries
-                switch (this.expectedHeader[idx]) {
-                    case 'submitters':
-                        newMotion.csvSubmitters = this.getSubmitters(line[idx]);
-                        break;
-                    case 'category':
-                        newMotion.csvCategory = this.getCategory(line[idx]);
-                        break;
-                    case 'motion block':
-                        newMotion.csvMotionblock = this.getMotionBlock(line[idx]);
-                        break;
-                    default:
-                        newMotion.motion[this.expectedHeader[idx]] = line[idx];
-                }
-            }
-            const updateModels = this.getDuplicates(newMotion.motion);
-            if (updateModels.length) {
-                this.setError(newMotion, 'Duplicates');
-            }
-            this._entries.push({ newMotion: newMotion, duplicates: updateModels });
-        });
-        this.newEntries.next(this._entries);
-        this.updatePreview();
     }
 
     /**
@@ -253,80 +128,32 @@ export class MotionImportService {
         this.newCategories = await this.createNewCategories();
         this.newSubmitters = await this.createNewUsers();
 
-        for (const entry of this._entries) {
-            if (entry.newMotion.status !== 'new') {
+        for (const entry of this.entries) {
+            if (entry.status !== 'new') {
                 continue;
             }
-            const openBlocks = entry.newMotion.solveMotionBlocks(this.newMotionBlocks);
+            const openBlocks = (entry.newEntry as ViewCsvCreateMotion).solveMotionBlocks(this.newMotionBlocks);
             if (openBlocks) {
-                this.setError(entry.newMotion, 'MotionBlock');
-                // TODO error handling if not all submitters could be matched
+                this.setError(entry, 'MotionBlock');
                 this.updatePreview();
                 continue;
             }
-            const openCategories = entry.newMotion.solveCategory(this.newCategories);
+            const openCategories = (entry.newEntry as ViewCsvCreateMotion).solveCategory(this.newCategories);
             if (openCategories) {
-                this.setError(entry.newMotion, 'Category');
+                this.setError(entry, 'Category');
                 this.updatePreview();
                 continue;
             }
-            const openUsers = entry.newMotion.solveSubmitters(this.newSubmitters);
+            const openUsers = (entry.newEntry as ViewCsvCreateMotion).solveSubmitters(this.newSubmitters);
             if (openUsers) {
-                this.setError(entry.newMotion, 'Submitters');
+                this.setError(entry, 'Submitters');
                 this.updatePreview();
                 continue;
             }
-            await this.repo.create(entry.newMotion.motion);
-            entry.newMotion.done();
+            await this.repo.create((entry.newEntry as ViewCsvCreateMotion).motion);
+            entry.status = 'done';
         }
         this.updatePreview();
-    }
-
-    /**
-     * Checks the dataStore for duplicates
-     * @returns an array of duplicates with the same identifier.
-     * @param motion
-     */
-    public getDuplicates(motion: CreateMotion): ViewMotion[] {
-        return this.repo.getMotionDuplicates(motion);
-    }
-
-    /**
-     * counts the amount of duplicates that have no decision on the action to
-     * be taken
-     */
-    public updatePreview(): void {
-        const summary = {
-            total: 0,
-            new: 0,
-            duplicates: 0,
-            errors: 0,
-            done: 0
-        };
-        this._entries.forEach(entry => {
-            summary.total += 1;
-            if (entry.newMotion.status === 'done') {
-                summary.done += 1;
-                return;
-            } else if (entry.newMotion.status === 'error' && !entry.duplicates.length) {
-                // errors that are not due to duplicates
-                summary.errors += 1;
-                return;
-            } else if (entry.duplicates.length) {
-                summary.duplicates += 1;
-                return;
-            } else if (entry.newMotion.status === 'new') {
-                summary.new += 1;
-            }
-        });
-        this._preview = summary;
-    }
-
-    /**
-     * returns a subscribable representation of the new Users to be imported
-     */
-    public getNewEntries(): Observable<NewMotionEntry[]> {
-        return this.newEntries.asObservable();
     }
 
     /**
@@ -479,101 +306,6 @@ export class MotionImportService {
             );
         }
         return await Promise.all(promises);
-    }
-
-    /**
-     * Handler after a file was selected. Basic checking for type, then hand
-     * over to parsing
-     *
-     * @param event type is Event, but has target.files, which typescript doesn't seem to recognize
-     */
-    public onSelectFile(event: any): void {
-        // TODO type
-        if (event.target.files && event.target.files.length === 1) {
-            if (event.target.files[0].type === 'text/csv') {
-                this._rawFile = event.target.files[0];
-                this.readFile(event.target.files[0]);
-            } else {
-                this.matSnackbar.open(this.translate.instant('Wrong file type detected. Import failed.'), '', {
-                    duration: 3000
-                });
-                this.clearPreview();
-                this._rawFile = null;
-            }
-        }
-    }
-
-    /**
-     * Rereads the (previously selected) file, if present. Thought to be triggered
-     * by parameter changes on encoding, column, text separators
-     */
-    public refreshFile(): void {
-        if (this._rawFile) {
-            this.readFile(this._rawFile);
-        }
-    }
-
-    /**
-     * (re)-reads a given file with the current parameter
-     */
-    private readFile(file: File): void {
-        this.reader.readAsText(file, this.encoding);
-    }
-
-    /**
-     * Checks the first line of the csv (the header) for consistency (length)
-     * @param row expected to be an array parsed from the first line of a csv file
-     */
-    private checkHeader(row: string[]): boolean {
-        const snackbarDuration = 3000;
-        if (row.length < 4) {
-            this.matSnackbar.open(this.translate.instant('The file has too few columns to be parsed properly.'), '', {
-                duration: snackbarDuration
-            });
-
-            this.clearPreview();
-            return false;
-        } else if (row.length < this.expectedHeader.length) {
-            this.matSnackbar.open(
-                this.translate.instant('The file seems to have some ommitted columns. They will be considered empty.'),
-                '',
-                { duration: snackbarDuration }
-            );
-        } else if (row.length > this.expectedHeader.length) {
-            this.matSnackbar.open(
-                this.translate.instant('The file seems to have additional columns. They will be ignored.'),
-                '',
-                { duration: snackbarDuration }
-            );
-        }
-        return true;
-    }
-
-    /**
-     * Resets the data and preview (triggered upon selecting an invalid file)
-     */
-    public clearPreview(): void {
-        this._entries = [];
-        this.newEntries.next([]);
-        this._preview = null;
-    }
-
-    /**
-     * set a list of short names for error, indicating which column failed
-     */
-    public setError(motion: ViewCsvCreateMotion, error: string): void {
-        if (errorList.hasOwnProperty(error) && !motion.errors.includes(error)) {
-            motion.errors.push(error);
-            motion.status = 'error';
-        }
-    }
-
-    /**
-     * Get an extended error description.
-     * @param error
-     */
-    public verbose(error: string): string {
-        return errorList[error];
     }
 
     /**
